Gaussian filtering function for 1D arrays
Due to the intrinsic limit in gaussian filtering process, the array is extended from the original size n to n + 2(r-1), with 2 additional parts connected to the start and the end evenly. The values of elements in two additional parts are selected from the start and the end of the original array, respectively.